refactored: songs come in mainplay&send to sidebar
